feat(mapview): remove componentDidUpdate logic
This logic was added all the way back in e2ff88dcfc9ee74c8c627030ba96adb3582dc0ba
where the region prop was explicitly set to null after spreading the rest of the props.
This is no longer the case, so the code is redundant, at best, as the region is also set
by spreading the props.